refactor: Major re-organisation of plugin backend
* Merge client and dashboard packages into one. Move all the dashboard related logic into dashboard package

* Report package deals with all reporting related logic and generates PDF reports

* Check permissions on dashboard and its folder to support inherited permissions

* Fix minor bug in panels filtering
