+++ pkg/plugin/chrome/types.go
@@ -16,7 +16,7 @@ type PDFOptions struct {
 
 // Instance is the interface remote and local chrome must implement.
 type Instance interface {
-	NewTab(logger log.Logger, conf config.Config) *Tab
+	NewTab(logger log.Logger, conf *config.Config) *Tab
 	Name() string
 	Close(logger log.Logger)
 }
