--- pkg/plugin/report/errors.go
@@ -1,5 +1,0 @@
-package report
-
-import "errors"
-
-var ErrEmptyDashboard = errors.New("empty dashboard model")
